@@ -14,7 +14,7 @@
     "vendor"
   ],
   "types": "bson.d.ts",
-  "version": "6.2.0",
+  "version": "6.10.2",
   "author": {
     "name": "The MongoDB NodeJS Team",
     "email": "dbx-node@mongodb.com"
@@ -27,41 +27,40 @@
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.36.4",
-    "@octokit/core": "^4.2.4",
-    "@rollup/plugin-node-resolve": "^15.1.0",
-    "@rollup/plugin-typescript": "^11.1.2",
-    "@types/chai": "^4.3.5",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.17.3",
-    "@types/sinon": "^10.0.16",
-    "@types/sinon-chai": "^3.2.9",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@microsoft/api-extractor": "^7.47.5",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/chai": "^4.3.17",
+    "@types/mocha": "^10.0.7",
+    "@types/node": "^22.1.0",
+    "@types/sinon": "^17.0.3",
+    "@types/sinon-chai": "^3.2.12",
+    "@typescript-eslint/eslint-plugin": "^8.0.1",
+    "@typescript-eslint/parser": "^8.0.1",
     "benchmark": "^2.1.4",
-    "chai": "^4.3.7",
+    "chai": "^4.4.1",
     "chalk": "^5.3.0",
-    "eslint": "^8.46.0",
-    "eslint-config-prettier": "^8.10.0",
+    "dbx-js-tools": "github:mongodb-js/dbx-js-tools",
+    "eslint": "^9.8.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-no-bigint-usage": "file:etc/eslint/no-bigint-usage",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-tsdoc": "^0.2.17",
-    "magic-string": "^0.30.2",
-    "mocha": "10.2.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-tsdoc": "^0.3.0",
+    "magic-string": "^0.30.11",
+    "mocha": "^10.7.0",
     "node-fetch": "^3.3.2",
     "nyc": "^15.1.0",
-    "prettier": "^2.8.8",
-    "rollup": "^3.27.2",
-    "sinon": "^15.2.0",
+    "prettier": "^3.3.3",
+    "rollup": "^4.20.0",
+    "sinon": "^18.0.0",
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.21",
-    "standard-version": "^9.5.0",
-    "tar": "^6.1.15",
-    "ts-node": "^10.9.1",
-    "tsd": "^0.28.1",
-    "typescript": "^5.0.4",
+    "tar": "^7.4.3",
+    "ts-node": "^10.9.2",
+    "tsd": "^0.31.1",
+    "typescript": "^5.5.4",
     "typescript-cached-transpile": "0.0.6",
-    "uuid": "^9.0.0"
+    "uuid": "^10.0.0"
   },
   "tsd": {
     "directory": "test/types",
@@ -103,12 +102,15 @@
     "check:tsd": "npm run build:dts && tsd",
     "check:web": "WEB=true mocha test/node",
     "check:web-no-bigint": "WEB=true NO_BIGINT=true mocha test/node",
-    "check:bench": "cd test/bench && npx tsc && node ./lib/index.js && mv benchmarks.json ../../.",
+    "check:granular-bench": "npm run build:bench && node ./test/bench/etc/run_granular_benchmarks.js",
+    "check:spec-bench": "npm run build:bench && node ./test/bench/lib/spec/bsonBench.js",
+    "check:custom-bench": "npm run build && node ./test/bench/custom/main.mjs",
+    "build:bench": "cd test/bench && npx tsc",
     "build:ts": "node ./node_modules/typescript/bin/tsc",
     "build:dts": "npm run build:ts && api-extractor run --typescript-compiler-folder node_modules/typescript --local && node etc/clean_definition_files.cjs",
     "build:bundle": "rollup -c rollup.config.mjs",
     "build": "npm run build:dts && npm run build:bundle",
-    "check:lint": "eslint -v && eslint --ext '.js,.ts' --max-warnings=0 src test && npm run build:dts && npm run check:tsd",
+    "check:lint": "ESLINT_USE_FLAT_CONFIG=false eslint -v && ESLINT_USE_FLAT_CONFIG=false eslint --ext '.js,.ts' --max-warnings=0 src test && npm run build:dts && npm run check:tsd",
     "format": "eslint --ext '.js,.ts' src test --fix",
     "check:coverage": "nyc --check-coverage npm run check:node",
     "prepare": "node etc/prepare.js",
