@@ -1,23 +1,23 @@
-const jwt = require('jsonwebtoken');
-const config = require('config');
+import jwt from 'jsonwebtoken';
+import config from 'config';
 
-module.exports = function(req, res, next) {
-    //Get token from header
+const auth = (req, res, next) => {
+    // Get token from header
     const token = req.header('x-auth-token');
 
-    //Check if no token
+    // Check if no token
     if (!token) {
-        return res.status(401).json({ msg: 'No token, authorization denied'});
-
+        return res.status(401).json({ msg: 'No token, authorization denied' });
     }
 
-    //Verify token
+    // Verify token
     try {
         const decoded = jwt.verify(token, config.get('jwtSecret'));
-
         req.user = decoded.user;
         next();
     } catch (err) {
         res.status(401).json({ msg: 'Token is not valid' });
     }
 };
+
+export default auth;
