@@ -1,5 +1,6 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
+import mongoose from 'mongoose';
+
+const { Schema } = mongoose;
 
 const PostSchema = new Schema({
     user: {
@@ -10,12 +11,8 @@ const PostSchema = new Schema({
         type: String,
         required: true
     },
-    name: {
-        type: String
-    },
-    avatar: {
-        type: String
-    },
+    name: String,
+    avatar: String,
     likes: [
         {
             user: {
@@ -34,12 +31,8 @@ const PostSchema = new Schema({
                 type: String,
                 required: true
             },
-            name: {
-                type: String
-            },
-            avatar: {
-                type: String
-            },
+            name: String,
+            avatar: String,
             date: {
                 type: Date,
                 default: Date.now
@@ -52,4 +45,6 @@ const PostSchema = new Schema({
     }
 });
 
-module.exports = Post = mongoose.model('post', PostSchema);
+const Post = mongoose.model('post', PostSchema);
+
+export default Post;
