@@ -1,12 +1,12 @@
-const express = require('express');
-const router = express.Router();
-const bcrypt = require('bcryptjs');
-const auth = require('../../middleware/auth');
-const jwt = require('jsonwebtoken');
-const config = require('config');
-const { body, validationResult } = require('express-validator');
+import express from 'express';
+import bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+import config from 'config';
+import { body, validationResult } from 'express-validator';
+import auth from '../../middleware/auth.js';
+import User from '../../models/Users.js';
 
-const User = require('../../models/Users');
+const router = express.Router();
 
 // @route  GET api/auth
 // @desc   Test route
@@ -21,55 +21,58 @@ router.get('/', auth, async (req, res) => {
     }
 });
 
-// @route  POST api/users
+// @route  POST api/auth
 // @desc   Authenticate user & get token
 // @access Public
-router.post('/', [
-    body('email', 'Please include a valid email').isEmail(),
-    body('password', 'Password is required').exists()
-],
-async (req, res) => { 
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-    }    
-    
-    const { email, password } =req.body;
-
-    try {
-//See if user exists
-        let user = await User.findOne({ email });
-
-        if (!user) {
-            return res. status(400).json({ errors: [{ msg: 'Invalid Credentials' }] });
+router.post(
+    '/',
+    [
+        body('email', 'Please include a valid email').isEmail(),
+        body('password', 'Password is required').exists()
+    ],
+    async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
         }
 
-        const isMatch = await bcrypt.compare(password, user.password);
+        const { email, password } = req.body;
 
-        if (!isMatch) {
-            return res.status(400).json({ errors: [{ msg: 'Invalid credentials' }] });
-        }
-// Return jsonwebtoken
-        const payload = {
-            user: {
-                id: user.id
+        try {
+            // See if user exists
+            let user = await User.findOne({ email });
+
+            if (!user) {
+                return res.status(400).json({ errors: [{ msg: 'Invalid Credentials' }] });
             }
+
+            const isMatch = await bcrypt.compare(password, user.password);
+
+            if (!isMatch) {
+                return res.status(400).json({ errors: [{ msg: 'Invalid credentials' }] });
+            }
+
+            // Return jsonwebtoken
+            const payload = {
+                user: {
+                    id: user.id
+                }
+            };
+
+            jwt.sign(
+                payload,
+                config.get('jwtSecret'),
+                { expiresIn: 360000 },
+                (err, token) => {
+                    if (err) throw err;
+                    res.json({ token });
+                }
+            );
+        } catch (err) {
+            console.error(err.message);
+            res.status(500).send('Server error');
         }
-
-        jwt.sign(
-            payload,
-            config.get('jwtSecret'),
-            { expiresIn: 360000 },
-            (err, token) => {
-                if(err) throw err;
-                res.json({ token });
-            });
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server error');    
     }
+);
 
-
-});
-
-module.exports = router;
+export default router;
