@@ -1,220 +1,202 @@
-const express = require('express');
-const router = express.Router();
-const { body, validationResult } = require('express-validator');
-const auth = require('../../middleware/auth');
+import express from 'express';
+import { body, validationResult } from 'express-validator';
+import auth from '../../middleware/auth.js';
+import Post from '../../models/Post.js';
+import Profile from '../../models/Profile.js';
+import User from '../../models/Users.js';
 
-const Post = require('../../models/Post');
-const Profile = require('../../models/Profile');
-const User = require('../../models/Users');
+const router = express.Router();
 
 // @route  POST api/posts
 // @desc   Create a post
 // @access Private
-router.post('/', [auth, [
-    body('text', 'Text is required').not().isEmpty()
-    ] 
-    ],
-    async (req, res) => {
-        const errors = validationResult(req);
-        if (!errors.isEmpty()) {
-            return res.status(400).json({ errors: errors.array() });
-        }
+router.post(
+  '/',
+  [auth, [body('text', 'Text is required').not().isEmpty()]],
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
     try {
-        const user = await User.findById(req.user.id).select('-password');
+      const user = await User.findById(req.user.id).select('-password');
 
-        const newPost = new Post({
-            text: req.body.text,
-            name: user.name,
-            avatar: user.avatar,
-            user: req.user.id
-        });
+      const newPost = new Post({
+        text: req.body.text,
+        name: user.name,
+        avatar: user.avatar,
+        user: req.user.id,
+      });
 
-        const post = await newPost.save();
-
-        res.json(post);
+      const post = await newPost.save();
+      res.json(post);
     } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server Error');
+      console.error(err.message);
+      res.status(500).send('Server Error');
     }
-    }
+  }
 );
 
 // @route  GET api/posts
 // @desc   Get all posts
 // @access Private
 router.get('/', auth, async (req, res) => {
-    try {
-        const posts = await Post.find().sort({ date: -1 });
-
-        res.json(posts);
-    } catch (err) {
-       console.error(err.message);
-       res.status(500).send('Server Error');
-    }
-        
+  try {
+    const posts = await Post.find().sort({ date: -1 });
+    res.json(posts);
+  } catch (err) {
+    console.error(err.message);
+    res.status(500).send('Server Error');
+  }
 });
 
 // @route  GET api/posts/:id
 // @desc   Get post by ID
 // @access Private
 router.get('/:id', auth, async (req, res) => {
-    try {
-        const post = await Post.findById(req.params.id);
+  try {
+    const post = await Post.findById(req.params.id);
 
-        if (!post) {
-            return res.status(404).json({ msg: 'Post not found' });
-        }
+    if (!post) {
+      return res.status(404).json({ msg: 'Post not found' });
+    }
 
-        res.json(post);
-    } catch (err) {
-       console.error(err.message);
-       if (err.kind === 'ObjectId') {
-        return res.status(404).json({ msg: 'Post not found' });
+    res.json(post);
+  } catch (err) {
+    console.error(err.message);
+    if (err.kind === 'ObjectId') {
+      return res.status(404).json({ msg: 'Post not found' });
     }
-       res.status(500).send('Server Error');
-    }
-        
+    res.status(500).send('Server Error');
+  }
 });
 
 // @route  DELETE api/posts/:id
 // @desc   Delete a post
 // @access Private
 router.delete('/:id', auth, async (req, res) => {
-    try {
-        const post = await Post.findById(req.params.id);
+  try {
+    const post = await Post.findById(req.params.id);
 
-        if (!post) {
-            return res.status(404).json({ msg: 'Post not found' });
-        }
-
-        if (post.user.toString() !== req.user.id) {
-            return res.status(401).json({ msg: 'User not authorized' });
-        }
-
-        await post.deleteOne();
-
-        res.json({ msg: 'Post Removed' });
-    } catch (err) {
-        console.error(err.message);
-        if (err.kind === 'ObjectId') {
-            return res.status(404).json({ msg: 'Post not found' });
-        }
-        res.status(500).send('Server Error');
+    if (!post) {
+      return res.status(404).json({ msg: 'Post not found' });
     }
+
+    if (post.user.toString() !== req.user.id) {
+      return res.status(401).json({ msg: 'User not authorized' });
+    }
+
+    await post.deleteOne();
+    res.json({ msg: 'Post Removed' });
+  } catch (err) {
+    console.error(err.message);
+    if (err.kind === 'ObjectId') {
+      return res.status(404).json({ msg: 'Post not found' });
+    }
+    res.status(500).send('Server Error');
+  }
 });
 
-// @route  PUT api/posts/likes/:id
+// @route  PUT api/posts/like/:id
 // @desc   Like a post
 // @access Private
 router.put('/like/:id', auth, async (req, res) => {
-    try {
-        const post = await Post.findById(req.params.id);
+  try {
+    const post = await Post.findById(req.params.id);
 
-        if (post.likes.filter(like => like.user.toString() === req.user.id).length > 0) {
-            return res.status(400).json({ msg: 'Post already liked' });
-        }
-
-        post.likes.unshift({ user: req.user.id });
-
-        await post.save();
-
-        res.json(post.likes);
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server Error');
+    if (post.likes.some(like => like.user.toString() === req.user.id)) {
+      return res.status(400).json({ msg: 'Post already liked' });
     }
+
+    post.likes.unshift({ user: req.user.id });
+    await post.save();
+    res.json(post.likes);
+  } catch (err) {
+    console.error(err.message);
+    res.status(500).send('Server Error');
+  }
 });
 
-// @route  PUT api/posts/unlikes/:id
+// @route  PUT api/posts/unlike/:id
 // @desc   Unlike a post
 // @access Private
 router.put('/unlike/:id', auth, async (req, res) => {
-    try {
-        const post = await Post.findById(req.params.id);
+  try {
+    const post = await Post.findById(req.params.id);
 
-        //Check if post has already been liked
-        if (post.likes.filter(like => like.user.toString() === req.user.id).length === 0) {
-            return res.status(400).json({ msg: 'Post has not been liked' });
-        }
-
-        //Get remove index
-        const removeIndex = post.likes.map(like => like.user.toString()).indexOf(req.user.id);
-
-        post.likes.splice(removeIndex, 1);
-
-        await post.save();
-
-        res.json(post.likes);
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server Error');
+    if (!post.likes.some(like => like.user.toString() === req.user.id)) {
+      return res.status(400).json({ msg: 'Post has not been liked' });
     }
+
+    post.likes = post.likes.filter(like => like.user.toString() !== req.user.id);
+    await post.save();
+
+    res.json(post.likes);
+  } catch (err) {
+    console.error(err.message);
+    res.status(500).send('Server Error');
+  }
 });
 
 // @route  POST api/posts/comment/:id
 // @desc   Comment on a post
 // @access Private
-router.post('/comment/:id', [auth, [
-    body('text', 'Text is required').not().isEmpty()
-] ],
-    async (req, res) => {
-        const errors = validationResult(req);
-        if (!errors.isEmpty()) {
-            return res.status(400).json({ errors: errors.array() });
-        }
-
-        try {
-            const user = await User.findById(req.user.id).select('-password');
-            const post = await Post.findById(req.params.id);
-
-            const newComment = {
-                text: req.body.text,
-                name: user.name,
-                avatar: user.avatar,
-                user: req.user.id
-            };
-
-            post.comments.unshift(newComment);
-
-            await post.save();
-
-            res.json(post.comments);
-        } catch (err) {
-            console.error(err.message);
-            res.sendStatus(500).send('Server Error');
-        }
+router.post(
+  '/comment/:id',
+  [auth, [body('text', 'Text is required').not().isEmpty()]],
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
     }
+
+    try {
+      const user = await User.findById(req.user.id).select('-password');
+      const post = await Post.findById(req.params.id);
+
+      const newComment = {
+        text: req.body.text,
+        name: user.name,
+        avatar: user.avatar,
+        user: req.user.id,
+      };
+
+      post.comments.unshift(newComment);
+      await post.save();
+      res.json(post.comments);
+    } catch (err) {
+      console.error(err.message);
+      res.status(500).send('Server Error');
+    }
+  }
 );
 
 // @route  DELETE api/posts/comment/:id/:comment_id
 // @desc   Delete comment
 // @access Private
 router.delete('/comment/:id/:comment_id', auth, async (req, res) => {
-    try {
-        const post = await Post.findById(req.params.id);
+  try {
+    const post = await Post.findById(req.params.id);
 
-        //Pull out comment
-        const comment = post.comments.find(comment => comment.id === req.params.comment_id);
+    // Find comment
+    const comment = post.comments.find(comment => comment.id === req.params.comment_id);
 
-        if (!comment) {
-            return res.status(404).json({ msg: 'Comment does not exist' });
-        }
-
-        if (comment.user.toString() !== req.user.id) {
-            return res.status(401).json({ msg: 'User not authorized' });
-        }
-
-        //Get remove imdex
-        const removeIndex = post.comments.map(comment => comment.user.toString()).indexOf(req.user.id);
-
-        post.comments.splice(removeIndex, 1);
-
-        await post.save();
-
-        res.json(post.comments);
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server error');
+    if (!comment) {
+      return res.status(404).json({ msg: 'Comment does not exist' });
     }
+
+    if (comment.user.toString() !== req.user.id) {
+      return res.status(401).json({ msg: 'User not authorized' });
+    }
+
+    post.comments = post.comments.filter(comment => comment.id !== req.params.comment_id);
+    await post.save();
+
+    res.json(post.comments);
+  } catch (err) {
+    console.error(err.message);
+    res.status(500).send('Server error');
+  }
 });
-module.exports = router;
+
+export default router;
