@@ -1,25 +1,24 @@
-const express = require('express');
-const request = require('request');
-const config = require('config');
-const router = express.Router();
-const auth = require('../../middleware/auth');
-const { body, validationResult } = require('express-validator');
+import express from 'express';
+import request from 'request';
+import config from 'config';
+import { body, validationResult } from 'express-validator';
 
-const Profile = require('../../models/Profile');
-const User = require('../../models/Users');
-const Post = require('../../models/Post');
+import auth from '../../middleware/auth.js';
+import Profile from '../../models/Profile.js';
+import User from '../../models/Users.js';
+import Post from '../../models/Post.js';
+
+const router = express.Router();
 
 // @route  GET api/profile/me
-// @desc   Get current users profile
+// @desc   Get current user's profile
 // @access Private
 router.get('/me', auth, async (req, res) => {
     try {
-        const profile = await Profile.findOne({ 
-            user: req.user.id 
-        }).populate('user', ['name', 'avatar']);
-        
+        const profile = await Profile.findOne({ user: req.user.id }).populate('user', ['name', 'avatar']);
+
         if (!profile) {
-            return res.status(400).json({ msg: 'There is no profile for this user'});
+            return res.status(400).json({ msg: 'There is no profile for this user' });
         }
         res.json(profile);
     } catch (err) {
@@ -29,91 +28,83 @@ router.get('/me', auth, async (req, res) => {
 });
 
 // @route  POST api/profile
-// @desc   Create or update users profile
+// @desc   Create or update user's profile
 // @access Private
 router.post(
-    '/', 
-    [ auth, [
-        body('status', 'Status is required').not().isEmpty(),
-        body('skills', 'Skills is required').not().isEmpty()
-    ]],
-async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-    }
-    const {
-        company,
-        website,
-        location,
-        status,
-        skills,
-        bio,
-        githubusername,
-        experience,
-        education,
-        youtube,
-        facebook,
-        twitter,
-        instagram,
-        linkedin
-    } = req.body;
-//Build profile object
-    const profileFields = {}
-    profileFields.user = req.user.id;
-    if (company) profileFields.company = company;
-    if (website) profileFields.website = website;
-    if (location) profileFields.location = location;
-    if (status) profileFields.status = status;
-    if (bio) profileFields.bio = bio;
-    if (githubusername) profileFields.githubusername = githubusername;
-    if (experience) profileFields.experience = experience;
-    if (education) profileFields.education = education;
-    if (skills) { profileFields.skills = skills.split(',').map(skills => skills.trim()); 
-}
-
-//Build social object
-    profileFields.social = {}
-    if (youtube) profileFields.social.youtube = youtube;
-    if (twitter) profileFields.social.twitter = twitter;
-    if (facebook) profileFields.social.facebook = facebook;
-    if (instagram) profileFields.social.instagram = instagram;
-    if (linkedin) profileFields.social.linkedin = linkedin;
-
-    try {
-        let profile = await Profile.findOne({ user: req.user.id });
-
-        if(profile) {
-            //Update
-            profile = await Profile.findOneAndUpdate(
-                { user: req.user.id },
-                { $set: profileFields },
-                { new: true }
-            );
-
-            return res.json(profile);
+    '/',
+    [
+        auth,
+        [
+            body('status', 'Status is required').not().isEmpty(),
+            body('skills', 'Skills is required').not().isEmpty(),
+        ],
+    ],
+    async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
         }
 
-            //Create
-            profile = new Profile(profileFields);
+        const {
+            company,
+            website,
+            location,
+            status,
+            skills,
+            bio,
+            githubusername,
+            experience,
+            education,
+            youtube,
+            facebook,
+            twitter,
+            instagram,
+            linkedin,
+        } = req.body;
 
+        const profileFields = {
+            user: req.user.id,
+            company,
+            website,
+            location,
+            status,
+            bio,
+            githubusername,
+            experience,
+            education,
+            skills: skills ? skills.split(',').map((skill) => skill.trim()) : [],
+            social: { youtube, twitter, facebook, instagram, linkedin },
+        };
+
+        try {
+            let profile = await Profile.findOne({ user: req.user.id });
+
+            if (profile) {
+                profile = await Profile.findOneAndUpdate(
+                    { user: req.user.id },
+                    { $set: profileFields },
+                    { new: true }
+                );
+                return res.json(profile);
+            }
+
+            profile = new Profile(profileFields);
             await profile.save();
             res.json(profile);
-    
-    } catch(err) {
-        console.error(err.message);
-        res.status(500).send('Server Error');
-    }
+        } catch (err) {
+            console.error(err.message);
+            res.status(500).send('Server Error');
+        }
     }
 );
 
 // @route  GET api/profile
-// @desc   Get all users profile
+// @desc   Get all users' profiles
 // @access Public
 router.get('/', async (req, res) => {
     try {
-       const profiles = await Profile.find().populate('user', ['name', 'avatar']);
-       res.json(profiles); 
+        const profiles = await Profile.find().populate('user', ['name', 'avatar']);
+        res.json(profiles);
     } catch (err) {
         console.error(err.message);
         res.status(500).send('Server Error');
@@ -123,38 +114,31 @@ router.get('/', async (req, res) => {
 // @route  GET api/profile/user/:user_id
 // @desc   Get profile by user ID
 // @access Public
-router.get(
-    '/user/:user_id', 
-    async (req, res) => {
+router.get('/user/:user_id', async (req, res) => {
     try {
-       const profile = await Profile.findOne({ 
-            user: req.params.user_id 
-        }).populate('user', ['name', 'avatar']);
-       
-       if (!profile) return res.status(400).json({ msg: 'Profile not found'});
-       res.json(profile); 
+        const profile = await Profile.findOne({ user: req.params.user_id }).populate('user', ['name', 'avatar']);
+
+        if (!profile) return res.status(400).json({ msg: 'Profile not found' });
+        res.json(profile);
     } catch (err) {
         console.error(err.message);
-        if (err.kind == 'ObjectId') {
-            return res.status(400).json({ msg: 'Profile not found' });   
+        if (err.kind === 'ObjectId') {
+            return res.status(400).json({ msg: 'Profile not found' });
         }
         res.status(500).send('Server Error');
     }
 });
 
 // @route  DELETE api/profile
-// @desc   Delete users profile, user & posts
+// @desc   Delete user profile, user & posts
 // @access Private
 router.delete('/', auth, async (req, res) => {
     try {
-      // Remove user posts
-      await Post.deleteMany({ user: req.user.id });  
-      //Remove Profile
-      await Profile.findOneAndDelete({ user: req.user.id });
-      //Remove User
-      await User.findOneAndDelete({ _id: req.user.id });
-      
-      res.json({ msg: 'User Deleted' });
+        await Post.deleteMany({ user: req.user.id });
+        await Profile.findOneAndDelete({ user: req.user.id });
+        await User.findOneAndDelete({ _id: req.user.id });
+
+        res.json({ msg: 'User Deleted' });
     } catch (err) {
         console.error(err.message);
         res.status(500).send('Server Error');
@@ -164,99 +148,69 @@ router.delete('/', auth, async (req, res) => {
 // @route  PUT api/profile/experience
 // @desc   Add profile experience
 // @access Private
-router.put('/experience', [ auth, [
-    body('title', 'Title is required').not().isEmpty(),
-    body('company', 'Company is required').not().isEmpty(),
-    body('from', 'From date is required').not().isEmpty()
-] ],
-async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+router.put(
+    '/experience',
+    [
+        auth,
+        [
+            body('title', 'Title is required').not().isEmpty(),
+            body('company', 'Company is required').not().isEmpty(),
+            body('from', 'From date is required').not().isEmpty(),
+        ],
+    ],
+    async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
+
+        const newExp = req.body;
+
+        try {
+            const profile = await Profile.findOne({ user: req.user.id });
+            profile.experience.unshift(newExp);
+            await profile.save();
+            res.json(profile);
+        } catch (err) {
+            console.error(err.message);
+            res.status(500).send('Server Error');
+        }
     }
+);
 
-    const {
-        title,
-        company,
-        location,
-        from,
-        to,
-        current,
-        description
-    } = req.body;
-
-    const newExp = {
-        title,
-        company,
-        location,
-        from,
-        to,
-        current,
-        description
-    }
-
-    try {
-        const profile = await Profile.findOne({ user: req.user.id });
-
-        profile.experience.unshift(newExp);
-
-        await profile.save();
-
-        res.json(profile);
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server Error');
-    }
-});
-
-// @route  POST api/profile/education
+// @route  PUT api/profile/education
 // @desc   Add education to profile
 // @access Private
-router.put('/education', [ auth, [
-    body('school', 'Schoolis required').not().isEmpty(),
-    body('degree', 'Degree is required').not().isEmpty(),
-    body('fieldofstudy', 'Field of study is required').not().isEmpty(),
-    body('from', 'From date is required').not().isEmpty()
-] ],
-async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+router.put(
+    '/education',
+    [
+        auth,
+        [
+            body('school', 'School is required').not().isEmpty(),
+            body('degree', 'Degree is required').not().isEmpty(),
+            body('fieldofstudy', 'Field of study is required').not().isEmpty(),
+            body('from', 'From date is required').not().isEmpty(),
+        ],
+    ],
+    async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
+
+        const newEdu = req.body;
+
+        try {
+            const profile = await Profile.findOne({ user: req.user.id });
+            profile.education.unshift(newEdu);
+            await profile.save();
+            res.json(profile);
+        } catch (err) {
+            console.error(err.message);
+            res.status(500).send('Server Error');
+        }
     }
-
-    const {
-        school,
-        degree,
-        fieldofstudy,
-        from,
-        to,
-        current,
-        description
-    } = req.body;
-
-    const newEdu = {
-        school,
-        degree,
-        fieldofstudy,
-        from,
-        to,
-        current,
-        description
-    }
-
-    try {
-        const profile = await Profile.findOne({ user: req.user.id });
-
-        profile.education.unshift(newEdu);
-
-        await profile.save();
-
-        res.json(profile);
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server Error');
-    }
-});
+);
 
 // @route  DELETE api/profile/experience/:exp_id
 // @desc   Delete profile experience
@@ -264,14 +218,8 @@ async (req, res) => {
 router.delete('/experience/:exp_id', auth, async (req, res) => {
     try {
         const profile = await Profile.findOne({ user: req.user.id });
-
-        //Get remove index
-        const removeIndex = profile.experience.map(item => item.id).indexOf(req.params.exp_id);
-
-        profile.experience.splice(removeIndex, 1);
-
+        profile.experience = profile.experience.filter((exp) => exp.id !== req.params.exp_id);
         await profile.save();
-
         res.json(profile);
     } catch (err) {
         console.error(err.message);
@@ -285,14 +233,8 @@ router.delete('/experience/:exp_id', auth, async (req, res) => {
 router.delete('/education/:edu_id', auth, async (req, res) => {
     try {
         const profile = await Profile.findOne({ user: req.user.id });
-
-        //Get remove index
-        const removeIndex = profile.education.map(item => item.id).indexOf(req.params.edu_id);
-
-        profile.education.splice(removeIndex, 1);
-
+        profile.education = profile.education.filter((edu) => edu.id !== req.params.edu_id);
         await profile.save();
-
         res.json(profile);
     } catch (err) {
         console.error(err.message);
@@ -301,28 +243,29 @@ router.delete('/education/:edu_id', auth, async (req, res) => {
 });
 
 // @route  GET api/profile/github/:username
-// @desc   Get user repos from Github
+// @desc   Get user repos from GitHub
 // @access Public
 router.get('/github/:username', (req, res) => {
     try {
         const options = {
             uri: `https://api.github.com/users/${req.params.username}/repos?per_page=5&sort=created:asc&client_id=${config.get('githubClientId')}&client_secret=${config.get('githubSecret')}`,
             method: 'GET',
-            headers: { 'user-agent': 'node.js' }
+            headers: { 'user-agent': 'node.js' },
         };
-        request(options, (error, response, body) => {
-            if(error) console.error(error);
 
-            if(response.statusCode !== 200) {
-                res.status(404).json({ msg: 'No Github profile found' });
+        request(options, (error, response, body) => {
+            if (error) console.error(error);
+
+            if (response.statusCode !== 200) {
+                return res.status(404).json({ msg: 'No GitHub profile found' });
             }
 
             res.json(JSON.parse(body));
         });
     } catch (err) {
         console.error(err.message);
-        res.status(500).send('Server Error');  
+        res.status(500).send('Server Error');
     }
 });
 
-module.exports = router;
+export default router;
