@@ -1,78 +1,81 @@
-const express = require('express');
-const router = express.Router();
-const gravatar = require('gravatar');
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const config = require('config');
-const { body, validationResult } = require('express-validator');
+import express from 'express';
+import gravatar from 'gravatar';
+import bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+import config from '../../config/default.json' assert { type: 'json' };
+ // Ensure your config file is compatible with ES modules
+import { body, validationResult } from 'express-validator';
+import User from '../../models/Users.js';
 
-// Bring in user model
-const User = require('../../models/Users');
+const router = express.Router();
 
 // @route  POST api/users
 // @desc   Register user
 // @access Public
-router.post('/', [
-    body('name', 'Name is required').notEmpty(),
-    body('email', 'Please include a valid email').isEmail(),
-    body('password', 'Please enter a password with 6 or more characters'
-    ).isLength({ min: 6 })
-],
-async (req, res) => { 
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-    }    
-    
-    const { name, email, password } =req.body;
-
-    try {
-//See if user exists
-        let user = await User.findOne({ email });
-
-        if (user) {
-            return res. status(400).json({ errors: [{ msg: 'User already exists' }] });
+router.post(
+    '/',
+    [
+        body('name', 'Name is required').notEmpty(),
+        body('email', 'Please include a valid email').isEmail(),
+        body('password', 'Please enter a password with 6 or more characters').isLength({ min: 6 })
+    ],
+    async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
         }
-// Get users gravatar
-        const avatar = gravatar.url(email, {
-            s: '200',
-            r: 'pg',
-            d: 'mm'
-        })
 
-        user = new User({
-            name,
-            email,
-            avatar,
-            password
-        });
-// Encrypt password
-        const salt = await bcrypt.genSalt(10);
-        
-        user.password = await bcrypt.hash(password, salt);
+        const { name, email, password } = req.body;
 
-        await user.save();
-// Return jsonwebtoken
-        const payload = {
-            user: {
-                id: user.id
+        try {
+            // See if user exists
+            let user = await User.findOne({ email });
+
+            if (user) {
+                return res.status(400).json({ errors: [{ msg: 'User already exists' }] });
             }
-        }
 
-        jwt.sign(
-            payload,
-            config.get('jwtSecret'),
-            { expiresIn: 360000 },
-            (err, token) => {
-                if(err) throw err;
-                res.json({ token });
+            // Get user's gravatar
+            const avatar = gravatar.url(email, {
+                s: '200',
+                r: 'pg',
+                d: 'mm'
             });
-    } catch (err) {
-        console.error(err.message);
-        res.status(500).send('Server error');    
+
+            user = new User({
+                name,
+                email,
+                avatar,
+                password
+            });
+
+            // Encrypt password
+            const salt = await bcrypt.genSalt(10);
+            user.password = await bcrypt.hash(password, salt);
+
+            await user.save();
+
+            // Return JSON Web Token
+            const payload = {
+                user: {
+                    id: user.id
+                }
+            };
+
+            jwt.sign(
+                payload,
+                config.jwtSecret, // Updated to direct import
+                { expiresIn: 360000 },
+                (err, token) => {
+                    if (err) throw err;
+                    res.json({ token });
+                }
+            );
+        } catch (err) {
+            console.error(err.message);
+            res.status(500).send('Server error');
+        }
     }
+);
 
-
-});
-
-module.exports = router;
+export default router;
