@@ -1,5 +1,10 @@
-const express = require('express');
-const connectDB = require('./config/db');
+import express from 'express';
+import connectDB from './config/db.js';
+
+import usersRouter from './routes/api/users.js';
+import profileRouter from './routes/api/profile.js';
+import postsRouter from './routes/api/posts.js';
+import authRouter from './routes/api/auth.js';
 
 const app = express();
 
@@ -7,16 +12,16 @@ const app = express();
 connectDB();
 
 // Init Middleware
-app.use(express.json({ extended: false }));
+app.use(express.json());
 
 app.get('/', (req, res) => res.send('API Running'));
 
 // Define Routes
-app.use('/api/users', require('./routes/api/users'));
-app.use('/api/profile', require('./routes/api/profile'));
-app.use('/api/posts', require('./routes/api/posts'));
-app.use('/api/auth', require('./routes/api/auth'));
+app.use('/api/users', usersRouter);
+app.use('/api/profile', profileRouter);
+app.use('/api/posts', postsRouter);
+app.use('/api/auth', authRouter);
 
-const PORT = process.env.PORT || 5000;
+const PORT = process.env.PORT || 5003;
 
 app.listen(PORT, () => console.log(`Server started on port ${PORT}`));
